refactor: 팀 점유율 관련 데이터 model 변경
