+++ android/app/src/main/java/com/yagubogu/data/dto/response/TeamOccupancyRatesResponse.kt
@@ -14,9 +14,7 @@ data class TeamOccupancyRatesResponse(
     fun toPresentation(): TeamOccupancyRates =
         TeamOccupancyRates(
             stadiumName = stadiumShortName,
-            rates =
-                teams.map { teamOccupancyRateDto: TeamOccupancyRateDto ->
-                    teamOccupancyRateDto.toPresentation()
-                },
+            awayTeamOccupancyRate = teams[0].toPresentation(),
+            homeTeamOccupancyRate = teams[1].toPresentation(),
         )
 }
