+++ android/app/src/main/java/com/yagubogu/presentation/home/model/StadiumStatsUiModel.kt
@@ -3,26 +3,28 @@ package com.yagubogu.presentation.home.model
 import java.time.LocalTime
 
 data class StadiumStatsUiModel(
-    val stadiumName: String,
-    val teamOccupancyStatuses: List<TeamOccupancyStatus>,
+    val stadiumOccupancyRates: List<TeamOccupancyRates>,
+//    val awayTeamStatus: TeamOccupancyStatus,
+//    val homeTeamStatus: TeamOccupancyStatus,
+//    val teamOccupancyStatuses: List<TeamOccupancyStatus>,
     val refreshTime: LocalTime = LocalTime.now(),
 ) {
-    val awayTeamStatus: TeamOccupancyStatus = teamOccupancyStatuses[FIRST_TEAM_INDEX]
-    val homeTeamStatus: TeamOccupancyStatus = teamOccupancyStatuses[SECOND_TEAM_INDEX]
+//    val awayTeamStatus: TeamOccupancyStatus = teamOccupancyStatuses[FIRST_TEAM_INDEX]
+//    val homeTeamStatus: TeamOccupancyStatus = teamOccupancyStatuses[SECOND_TEAM_INDEX]
 
-    val awayTeamBias: Double = remapRange(awayTeamStatus.percentage)
-    val homeTeamBias: Double = remapRange(homeTeamStatus.percentage)
+//    val awayTeamBias: Double = remapRange(awayTeamStatus.percentage)
+//    val homeTeamBias: Double = remapRange(homeTeamStatus.percentage)
 
-    private fun remapRange(value: Double): Double {
-        val scalingFactor: Double = (FULL_PERCENTAGE - CHART_END_PADDING_SIZE * 2) / FULL_PERCENTAGE
-        val percentResult: Double = CHART_END_PADDING_SIZE + value * scalingFactor
-        return percentResult / FULL_PERCENTAGE
-    }
+//    private fun remapRange(value: Double): Double {
+//        val scalingFactor: Double = (FULL_PERCENTAGE - CHART_END_PADDING_SIZE * 2) / FULL_PERCENTAGE
+//        val percentResult: Double = CHART_END_PADDING_SIZE + value * scalingFactor
+//        return percentResult / FULL_PERCENTAGE
+//    }
 
-    companion object {
-        private const val FULL_PERCENTAGE = 100.0
-        private const val CHART_END_PADDING_SIZE = 28.0
-        private const val FIRST_TEAM_INDEX = 0
-        private const val SECOND_TEAM_INDEX = 1
-    }
+//    companion object {
+//        private const val FULL_PERCENTAGE = 100.0
+//        private const val CHART_END_PADDING_SIZE = 28.0
+//        private const val FIRST_TEAM_INDEX = 0
+//        private const val SECOND_TEAM_INDEX = 1
+//    }
 }
