+++ android/app/src/main/java/com/yagubogu/presentation/home/model/TeamOccupancyRate.kt
@@ -1,8 +1,15 @@
 package com.yagubogu.presentation.home.model
 
+import androidx.annotation.ColorRes
 import com.yagubogu.domain.model.Team
+import com.yagubogu.presentation.util.getTeamColor
 
 data class TeamOccupancyRate(
     val team: Team,
     val occupancyRate: Double,
-)
+) {
+    @ColorRes
+    val teamColor: Int = team.getTeamColor()
+
+    val teamName: String = team.shortName
+}
