+++ android/app/src/main/res/layout/fragment_home.xml
@@ -218,7 +218,7 @@
                     android:layout_margin="20dp"
                     app:layout_constraintBottom_toTopOf="@id/constraint_show_more"
                     app:layout_constraintTop_toBottomOf="@id/tv_team_occupancy_title"
-                    app:stadiumStatsUiModel="@{stadiumStatsUiModel}" />
+                    app:teamOccupancyRates="@{stadiumStatsUiModel.stadiumOccupancyRates.first}" />
 
                 <androidx.constraintlayout.widget.ConstraintLayout
                     android:id="@+id/constraint_show_more"
